refactor(upload > watermark): move module at top level
