+++ src/admin/upload/watermark/canvas.js
